backend/requirements.txt
fastapi
firebase-admin
pillow
pycryptodome
pydantic[email]
python-dateutil
python-dotenv
uvicorn